+++ CHANGELOG.md
@@ -22,6 +22,11 @@
   fixes a bug in `MapReduceJob`'s chunkification when `n_runs` >= `n_jobs`,
   and defines a sequential parallel backend to run all jobs in the current thread
   [PR #232](https://github.com/appliedAI-Initiative/pyDVL/pull/232)
+- **New method**: Implements exact and monte carlo Least Core for data valuation,
+  adds `from_arrays()` class method to the `Dataset` and `GroupedDataset` classes,
+  adds `extra_values` argument to `ValuationResult`,
+  adds `compute_removal_score()` and `compute_random_removal_score()` helper functions
+  [PR #237](https://github.com/appliedAI-Initiative/pyDVL/pull/237)
 
 ## 0.3.0 - 💥 Breaking changes